requirements.txt
py-cord
discord-ext-prettyhelp
python-dotenv
sqlalchemy[asyncio]
requests
aiosqlite